perf(lcc): pack auction state and record fills in the linked library
AuctionState packs its four counters into two slots, and takeAuction
records fills through a linked-library entrypoint that computes the
award and updates the packed counters in the same delegatecall, so a
first fill costs one dirty write instead of a cold zero-to-nonzero
store: takeAuction drops ~19.3k gas (-7.7%). getAuctionState narrows
two member types with an identical wire encoding, and both counters
stay SafeCast-checked at the storage boundary.

+++ test/forge/lcc/LCCAuctionLib.t.sol
@@ -105,7 +105,7 @@ contract LCCAuctionLibTest is Test {
             uint256 award = LCCAuctionLib.fillAward(state, fill, offered, type(uint256).max);
 
             state.filledAmount += uint128(fill);
-            state.marginAwarded += award;
+            state.marginAwarded += uint128(award);
             totalAward += award;
         }
 
@@ -141,7 +141,7 @@ contract LCCAuctionLibTest is Test {
         uint256 price
     ) public pure {
         shortfall = uint128(bound(shortfall, 1, type(uint128).max));
-        uint256 marginAwarded = bound(uint256(awardedSeed), 0, pool);
+        uint128 marginAwarded = uint128(bound(uint256(awardedSeed), 0, pool));
         uint256 fillAmount = bound(uint256(fillSeed), 0, shortfall);
         stepDuration = uint32(bound(stepDuration, 1, type(uint32).max));
         decayBps = bound(decayBps, 1, BPS);
